requirements.txt
alpaca-py>=0.20.0
flask>=3.0.0
gunicorn>=21.0.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pytz>=2024.1
